# Conflicts:
#	base-core/core-metadata/src/main/java/io/dexterity/service/MetaDataService.java
#	base-core/core-metadata/src/main/java/io/dexterity/service/impl/LmdbMetaDataService.java
#	base-web/src/main/resources/application.yml

+++ base-common/src/main/java/io/dexterity/entity/LMDBEnvSettings.java
@@ -8,7 +8,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @Data
 public class LMDBEnvSettings {
-    private int maxSize;
+    private Long maxSize;
     private int maxDBInstance;
     private int maxReaders;
     private String filePosition;
